cleanReadings now checks all readings to make sure they have valid sensor types

+++ smarthome/dragonfly/management/commands/cleanReadings.py
@@ -16,9 +16,29 @@ class Command(BaseCommand):
                 print 'Sensor not in whitelist: {}'.format(i.toDict()['name'])
                 print "would you like to delete it? (y,n)"
                 ans = raw_input('--> ')
-                print "you said: {}".format(ans)
                 if ans == "y":
                     print "deleting {}".format(i.toDict()['name'])
+                    i.delete()
+                elif ans == "n":
+                    print "skipping"
+                else:
+                    print "I didn't understand, quiting"
+                    break
+        print "Done checking sensors, begin readings check.(This may take a while"
+        readings = models.Reading.objects.all()
+        count = 0
+        for i in readings:
+            count += 1
+            if count % 1000 == 0:
+                print "readings checked: {}".format(count)
+            reading = i.toDict()
+            if reading['sensor']['name'] not in whitelist:
+                print "\n"
+                print json.dumps(i.toDict(), indent=2)
+                print " does not have a sensor that matches whitelist, delete? (y,n)"
+                ans = raw_input('--> ')
+                if ans == "y":
+                    print "deleting {}".format(i.toDict())
                     i.delete()
                 elif ans == "n":
                     print "skipping"
